Sharing Source
Uploading and sharing the project files for Heist in the Deep Mine, the game I made solo for the Extra Jam #5 by Extra Credits.

Included are low poly models I created myself during the jam.

Audio files were removed as I'm not aware if I possess the rights to redistribute them as assets.